+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add hidden classes for touch devices to utility component
+- Add clsBelow option for Sticky component (uk-sticky-below)
 
 ### Fixed